refactor: extract shared session-start helpers
