--- apps/desktop/src/pages/agents/agents-page-utils.ts
@@ -1,10 +1,0 @@
-export {
-  emptyDraftSelections,
-  extractCompletionTimestamp,
-  isSameSelection,
-  normalizeSelectionForCatalog,
-  parseTimestamp,
-  pickDefaultSelectionForCatalog,
-  toContextStorageKey,
-  toTabsStorageKey,
-} from "./agents-page-selection";
